Footer added 1. `components` mv to src/ 2. `global` layout has the dark mode
