+++ src/app/page.js
@@ -1,7 +1,6 @@
-"use client";
-import Content from "./components/Content";
-import Header from "./components/Header";
-import Footer from "./components/Footer";
+import Content from "@/components/Content";
+import Header from "@/components/Header";
+import Footer from '@/components/Footer';
 
 export default function Home() {
   return (
