+++ src/components/Content.jsx
@@ -1,5 +1,5 @@
 'use client';
-import styles from '../page.module.css'
+import styles from './page.module.css'
 import { useRef, useEffect } from 'react';
 import { ScrollTrigger } from 'gsap/ScrollTrigger';
 import gsap from 'gsap';
